Algorithm.ShapeConstraint

Description

This package provides support functions to evaluate different shape-constraints for symbolic regression using modal arithmetic.

See https://direct.mit.edu/evco/article/30/1/75/99840 for more details.

@article{kronberger2022shape,
  title={Shape-Constrained Symbolic Regression—Improving Extrapolation with Prior Knowledge},
  author={Kronberger, Gabriel and de Fran{c{c}}a, Fabricio Olivetti and Burlacu, Bogdan and Haider, Christian and Kommenda, Michael},
  journal={Evolutionary computation},
  volume={30},
  number={1},
  pages={75--98},
  year={2022},
  publisher={MIT Press One Rogers Street, Cambridge, MA 02142-1209, USA journals-info~…}
}

Documentation

getViolationFun :: Evaluator -> [Shape] -> Domains -> ConstraintFun

Returns a constraint function given an Evaluator a list of Shape and the Domains of the input variables.

data Evaluator

The Evaluator indicates the algorithm used to calculate the constraint:

InnerInterval can result in false positives (i.e., it says it is feasible but it is not)

OuterInterval can result in false negatives (i.e., it says it is infeasible but it is not)

Sampling can result in false positives and usually requires a lot of samples

Hybrid - not yet implemented

Bisection can improve the results of OuterInterval

Constructors

InnerInterval

OuterInterval

Sampling Int

Hybrid

Bisection Int

data Shape

Data type describing the type of constraint:

Range of the function

Range of a partial derivative

Monotonicity (non-increasing and non-decreasing)

Partial monotonicity

Inflection, Convex or Concave

The monotonicity constraints require the index of the variable, the ranges require a tuple of the minimum and maximum values, the partial monotonicity also requires the range of the domain that has this property, inflection, convexity and concavity requires the first and second index of the partial derivatives.

Constructors

Range (Double, Double)
f(x) in [a, b]

DiffRng Int (Double, Double)
d f(x) / dx in [a, b]

NonIncreasing Int
d f(x) / dx in [-inf, 0]

NonDecreasing Int
d f(x) / dx in [0, inf]

PartialNonIncreasing Int (Double, Double)
d f(x) / dx in [-inf, 0], a <= x <= b

PartialNonDecreasing Int (Double, Double)
d f(x) / dx in [0, inf], a <= x <= b

Inflection Int Int
d^2 f(x)/dx^2 == 0

Convex Int Int
d^2 f(x)/dx^2 in (0, Infinity)

Concave Int Int
d^2 f(x)/dx^2 in (-Infinitiy,0)

Orphan instances

OptIntPow (Vector Double)

Instance details

Methods

(^.) :: Vector Double -> Int -> Vector Double